+++ html.library/class_v_l_c_b_1_1_event_consumer_service_with_diagnostics.js
@@ -1,5 +1,5 @@
 var class_v_l_c_b_1_1_event_consumer_service_with_diagnostics =
 [
     [ "reportDiagnostics", "class_v_l_c_b_1_1_event_consumer_service_with_diagnostics.html#ab2ae84025dde7311ecf888640fe4d1f1", null ],
-    [ "reportAllDiagnostics", "class_v_l_c_b_1_1_event_consumer_service_with_diagnostics.html#a735d07e6ddcae79803a4dab0b0bb1c44", null ]
+    [ "getDiagnosticCount", "class_v_l_c_b_1_1_event_consumer_service_with_diagnostics.html#ad229601d23c5d2a3d463fa34f309341e", null ]
 ];