+++ html.library/class_v_l_c_b_1_1_event_producer_service_with_diagnostics.js
@@ -1,5 +1,5 @@
 var class_v_l_c_b_1_1_event_producer_service_with_diagnostics =
 [
     [ "reportDiagnostics", "class_v_l_c_b_1_1_event_producer_service_with_diagnostics.html#a3209a71ebd396539dea9c34b55ff2071", null ],
-    [ "reportAllDiagnostics", "class_v_l_c_b_1_1_event_producer_service_with_diagnostics.html#a18e784fda895a6b0dbaee53c5a072a42", null ]
+    [ "getDiagnosticCount", "class_v_l_c_b_1_1_event_producer_service_with_diagnostics.html#af7ed8637dfa68ab872fda932f83c07f8", null ]
 ];