+++ html.library/search/classes_c.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['transport_0',['Transport',['../class_v_l_c_b_1_1_transport.html',1,'VLCB']]]
+  ['task_0',['Task',['../class_v_l_c_b_1_1_timed_response_1_1_task.html',1,'VLCB::TimedResponse']]],
+  ['timedresponse_1',['TimedResponse',['../class_v_l_c_b_1_1_timed_response.html',1,'VLCB']]],
+  ['transport_2',['Transport',['../class_v_l_c_b_1_1_transport.html',1,'VLCB']]]
 ];