+++ html.library/search/enumvalues_b.js
@@ -69,5 +69,6 @@ var searchData=
   ['producer_5fev_5fhappening_66',['PRODUCER_EV_HAPPENING',['../vlcbdefs_8hpp.html#a2c53a2367445493162ee7a864ff47b97adee61b472fb8585aeaacd7f07fb65d2d',1,'vlcbdefs.hpp']]],
   ['producer_5fev_5fnot_5fspecified_67',['PRODUCER_EV_NOT_SPECIFIED',['../vlcbdefs_8hpp.html#a2c53a2367445493162ee7a864ff47b97ac1907ef66e447afaa0e52eb8a02d76b1',1,'vlcbdefs.hpp']]],
   ['producer_5fev_5fslots_68',['PRODUCER_EV_SLOTS',['../vlcbdefs_8hpp.html#a2c53a2367445493162ee7a864ff47b97a4f08811a16f8b3ddb74da09df55bd799',1,'vlcbdefs.hpp']]],
-  ['pulsing_69',['Pulsing',['../class_v_l_c_b_1_1_l_e_d.html#a42ae133cb00a6e9e2135cf56d7ae1816a3220dd5d6784e5317390de4cfada5ff7',1,'VLCB::LED']]]
+  ['progress_69',['PROGRESS',['../class_v_l_c_b_1_1_timed_response.html#a15c1994433c8b6bdd141a97075ea9ac8ac6fdc9a393c61a74e966e85a63877f2d',1,'VLCB::TimedResponse']]],
+  ['pulsing_70',['Pulsing',['../class_v_l_c_b_1_1_l_e_d.html#a42ae133cb00a6e9e2135cf56d7ae1816a3220dd5d6784e5317390de4cfada5ff7',1,'VLCB::LED']]]
 ];