+++ html.sketch/search/classes_8.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['send_5fcontext_5ft_0',['send_context_t',['../struct_v_l_c_b_1_1send__context__t.html',1,'VLCB']]],
-  ['service_1',['Service',['../class_v_l_c_b_1_1_service.html',1,'VLCB']]]
+  ['task_0',['Task',['../class_v_l_c_b_1_1_timed_response_1_1_task.html',1,'VLCB::TimedResponse']]],
+  ['timedresponse_1',['TimedResponse',['../class_v_l_c_b_1_1_timed_response.html',1,'VLCB']]]
 ];