+++ html.sketch/search/functions_1.js
@@ -1,7 +1,4 @@
 var searchData=
 [
-  ['can2515_0',['CAN2515',['../class_v_l_c_b_1_1_c_a_n2515.html#aa5ec5c93b8bcfa30167551bc7dda867d',1,'VLCB::CAN2515']]],
-  ['canservice_1',['CanService',['../class_v_l_c_b_1_1_can_service.html#a558d311e484d9d00476c2cf7ec4b3109',1,'VLCB::CanService']]],
-  ['checkstartupaction_2',['checkStartupAction',['../namespace_v_l_c_b.html#aca9610abdafcfe61167ac97a325bc858',1,'VLCB']]],
-  ['createeventatindex_3',['createEventAtIndex',['../namespace_v_l_c_b.html#a77277b82574203c049b2799a9971c308',1,'VLCB']]]
+  ['begin_0',['begin',['../classstd_1_1initializer__list.html#a091f20db46528a0c66d7711133f9b6bb',1,'std::initializer_list::begin()'],['../namespace_v_l_c_b.html#a9b7733390999c3dfd8e922f7397a3cc2',1,'VLCB::begin()'],['../namespacestd.html#a5b3743752103933dc8699124dbd618a6',1,'std::begin()']]]
 ];