+++ html.sketch/search/functions_3.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['end_0',['end',['../classstd_1_1initializer__list.html#a51d09e34c4c0130f18f9f98320915344',1,'std::initializer_list::end()'],['../namespacestd.html#a71dca3f09ec2536959d378fef972a803',1,'std::end()']]]
+  ['doeseventexistatindex_0',['doesEventExistAtIndex',['../namespace_v_l_c_b.html#a47e47eb14d8fb4b4d99a19bd3b01f68e',1,'VLCB']]]
 ];