+++ html.sketch/search/functions_6.js
@@ -1,5 +1,11 @@
 var searchData=
 [
-  ['initializer_5flist_0',['initializer_list',['../classstd_1_1initializer__list.html#a99fd02613715235bd6bc89ca4530d6fe',1,'std::initializer_list::initializer_list() noexcept'],['../classstd_1_1initializer__list.html#a9426dfd4cdb1b381a9d11676810afb03',1,'std::initializer_list::initializer_list(const E *a, size_t l)']]],
-  ['iseventindexvalid_1',['isEventIndexValid',['../namespace_v_l_c_b.html#aac4641296c4560e3b893be1e7e697927',1,'VLCB']]]
+  ['getcanid_0',['getCANID',['../namespace_v_l_c_b.html#afdbd716415513eecc98597480bd85daa',1,'VLCB']]],
+  ['getcurrentmode_1',['getCurrentMode',['../namespace_v_l_c_b.html#afc9d9eaacfa65438d13a45e95ad7eb8f',1,'VLCB']]],
+  ['getdiagnosticcount_2',['getDiagnosticCount',['../class_v_l_c_b_1_1_internal_diagnostics_service.html#a845a102cf9feb93bfd72cefd142761aa',1,'VLCB::InternalDiagnosticsService']]],
+  ['geteventevval_3',['getEventEVval',['../namespace_v_l_c_b.html#af4db790a620d538341751018e3fe7107',1,'VLCB']]],
+  ['getfreeeeprombase_4',['getFreeEEPROMbase',['../namespace_v_l_c_b.html#a5045aa3e373743aef424a957c0489671',1,'VLCB']]],
+  ['getnodenum_5',['getNodeNum',['../namespace_v_l_c_b.html#aecec03956461a72e544cfa0b2a945c3c',1,'VLCB']]],
+  ['getserviceid_6',['getServiceID',['../class_v_l_c_b_1_1_internal_diagnostics_service.html#ad84b11ea283e1d807d130943cee40810',1,'VLCB::InternalDiagnosticsService']]],
+  ['getserviceversionid_7',['getServiceVersionID',['../class_v_l_c_b_1_1_internal_diagnostics_service.html#a98fef7e4486654cb9ae6224e69423571',1,'VLCB::InternalDiagnosticsService']]]
 ];