+++ src/EventConsumerServiceWithDiagnostics.cpp
@@ -20,17 +20,13 @@ void EventConsumerServiceWithDiagnostics::reportDiagnostics(byte serviceIndex, b
       break;
     default:
       controller->sendGRSP(OPC_RDGN, serviceIndex, GRSP_INVALID_DIAGNOSTIC);
-      return;
+      break;
   }
 }
 
-void EventConsumerServiceWithDiagnostics::reportAllDiagnostics(byte serviceIndex)
+int EventConsumerServiceWithDiagnostics::getDiagnosticCount()
 {
-  byte diagCount = 2;
-  controller->sendDGN(serviceIndex, 0, diagCount);
-  for (byte i = 1; i <= diagCount ; ++i)
-  {
-    reportDiagnostics(serviceIndex, i);
-  }
-}      
+  return 2;
+}
+
 }