+++ src/EventProducerServiceWithDiagnostics.h
@@ -16,7 +16,7 @@ class EventProducerServiceWithDiagnostics : public EventProducerService
 {
 public:
   virtual void reportDiagnostics(byte serviceIndex, byte diagnosticsCode) override;
-  virtual void reportAllDiagnostics(byte serviceIndex) override;
+  virtual int getDiagnosticCount() override;
 
 };