+++ src/MinimumNodeService.cpp
@@ -7,6 +7,7 @@
 
 #include "MinimumNodeService.h"
 #include "Controller.h"
+#include "TimedResponse.h"
 
 namespace VLCB
 {
@@ -303,6 +304,21 @@ void MinimumNodeService::handleRequestNodeParameters()
   controller->messageActedOn();
 }
 
+class RespondParam : public TimedResponse::Task
+{
+public: 
+  RespondParam(Controller *controller) : Task(controller) {}
+  virtual TimedResponse::Result runStep() override
+  {
+    if (sequence > controller->getParam(PAR_NUM))
+    {
+      return TimedResponse::FINISHED;
+    }
+    controller->sendMessageWithNN(OPC_PARAN, sequence, controller->getParam((VlcbParams) sequence));
+    return TimedResponse::PROGRESS;
+  }
+};
+  
 void MinimumNodeService::handleRequestNodeParameter(const VlcbMessage *msg, unsigned int nn)
 {
   if (!isThisNodeNumber(nn))
@@ -322,10 +338,7 @@ void MinimumNodeService::handleRequestNodeParameter(const VlcbMessage *msg, unsi
 
     if ((paran == 0) && notFcuCompatible)
     {
-      for (byte i = 0; i <= controller->getParam(PAR_NUM); i++)
-      {
-        controller->sendMessageWithNN(OPC_PARAN, i, controller->getParam((VlcbParams) i));
-      }
+      controller->addTimedResponseTask(new RespondParam(controller));
     }
     else if (paran <= controller->getParam(PAR_NUM))
     {
@@ -382,6 +395,25 @@ static int countServices(const VLCB::ArrayHolder<Service *> &services)
   return count;
 }
 
+class RespondService : public TimedResponse::Task
+{
+public: 
+  RespondService(Controller *controller) : Task(controller) {}
+  virtual TimedResponse::Result runStep() override
+  {
+    if (sequence >= controller->getServices().size())
+    {
+      return TimedResponse::FINISHED;
+    }
+    Service * svc = controller->getServices()[sequence];
+    if (svc->getServiceID() > 0)
+    {
+      controller->sendMessageWithNN(OPC_SD, sequence+1, svc->getServiceID(), svc->getServiceVersionID());
+    }
+    return TimedResponse::PROGRESS;
+  }
+};
+
 void MinimumNodeService::handleRequestServiceDefinitions(const VlcbMessage *msg, unsigned int nn)
 {
   if (!isThisNodeNumber(nn))
@@ -403,16 +435,7 @@ void MinimumNodeService::handleRequestServiceDefinitions(const VlcbMessage *msg,
     controller->sendMessageWithNN(OPC_SD, 0, 0, serviceCount);
 
     // and then details of each service.
-    byte svcIndex = 0;
-    for (auto svc: controller->getServices())
-    {
-      ++svcIndex;
-      if (svc->getServiceID() > 0)
-      {
-        // TODO: Need to space out these messages, put in a queue or use a TimedResponse structure.
-        controller->sendMessageWithNN(OPC_SD, svcIndex, svc->getServiceID(), svc->getServiceVersionID());
-      }
-    }
+    controller->addTimedResponseTask(new RespondService(controller));
   }
   else if (serviceIndex <= controller->getServices().size())
   {